+++ src/util/clipping/clippingPlaneHelper.ts
@@ -18,7 +18,7 @@ import Feature from 'ol/Feature.js';
 import LineString from 'ol/geom/LineString.js';
 import { offset } from 'ol/sphere.js';
 import Polygon from 'ol/geom/Polygon.js';
-import { check, checkMaybe } from '@vcsuite/check';
+import { check, maybe, optional } from '@vcsuite/check';
 import Projection, {
   mercatorProjection,
   wgs84Projection,
@@ -159,7 +159,7 @@ export function createClippingPlaneCollection(
 ): ClippingPlaneCollection | null {
   check(feature, Feature);
   check(options, Object);
-  checkMaybe(transformMatrix, Matrix4);
+  check(transformMatrix, optional(Matrix4));
 
   const clippingPlanes = [];
   const geometry = feature.getGeometry() as Geometry;
@@ -457,7 +457,7 @@ export function getClippingOptions(
   feature?: Feature,
   infinite = false,
 ): ClippingPlaneCreationOptions {
-  checkMaybe(feature, Feature);
+  check(feature, maybe(Feature));
   check(infinite, Boolean);
 
   const vertical = feature